// src/Molecules/Components/Dropdown/type.ts
export type DropdownIcon = IconName | ImageIconConfig | React.ReactNode;

export interface DropdownOption {
  label: React.ReactNode;
  triggerLabel?: React.ReactNode;
  value: string;
  /**
   * @description The icon to include at start of the label. Can use a custom image using the ImageIconConfig (object with url and alt properties) type
   */
  icon?: DropdownIcon;
}
// ...
export enum DropdownCollapsibleOptionModes {
  MULTI_SELECT = "MULTI_SELECT",
  RADIO = "RADIO",
}

export interface DropdownCollapsibleInternalOptionValue {
  groupId: string;
  option: DropdownOption;
  mode?: DropdownCollapsibleOptionModes;
}
export interface DropdownCollapsibleOptionsGroup {
  label: React.ReactNode;
  options: DropdownOption[];
  mode: DropdownCollapsibleOptionModes;
  defaultOpen?: boolean;
}